requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image
langchain_community
textstat
rembg
langchain_google_genai
youtube_transcript_api
pathlib
openai
huggingface_hub
matplotlib
wordcloud
sentence-transformers
uvicorn
ctransformers
python-box
tensorflow
plotly 
openpyxl